Move navigation utils to util folder in check-in app root
