+++ src/applications/check-in/day-of/containers/withToken.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
 
 import { useSessionStorage } from '../../hooks/useSessionStorage';
-import { URLS } from '../utils/navigation';
+import { URLS } from '../../utils/navigation/day-of';
 import { useFormRouting } from '../../hooks/useFormRouting';
 import { makeSelectCheckInData } from '../hooks/selectors';
