+++ src/applications/check-in/pre-check-in/pages/Demographics/index.jsx
@@ -9,7 +9,7 @@ import BackButton from '../../components/BackButton';
 import DemographicsDisplay from '../../../components/pages/demographics/DemographicsDisplay';
 import recordEvent from 'platform/monitoring/record-event';
 import { recordAnswer } from '../../actions';
-import { URLS } from '../../utils/navigation';
+import { URLS } from '../../../utils/navigation/pre-check-in';
 
 import { makeSelectVeteranData } from '../../../selectors';
