+++ src/applications/check-in/utils/navigation/pre-check-in/navigation.unit.spec.js
